requirements.txt
langchain-community
langchain-unstructured 

pypandoc
langchain-core
langchain-milvus
scrapy
w3lib
pymilvus
torch
sentence-transformers
langchain-huggingface
python-magic
pdfminer.six
pi-heif
pdf2image
unstructured[all-docs]
langchain-milvus